[Fix] 재생성 위치 초기화

+++ Assets/Scripts/SOs/HumanSO.cs
@@ -9,6 +9,6 @@ public class HumanSO : ScriptableObject
     public float maxFatigueInflicted; // 몬스터의 피로도(체력) 최대 증가치
     public float cooldown;            // 공격 쿨타임
     public float speed;               // 속도
-    public float lifeInflicted;       // 스테이지의 health 감소치
-    public float coin;                // 처치 시 플레이어에 지급되는 재화량
+    public int lifeInflicted;         // 스테이지의 health 감소치
+    public int coin;                // 처치 시 플레이어에 지급되는 재화량
 }
